Fix build after r158967 https://bugs.webkit.org/show_bug.cgi?id=124160
Patch by Nick Diego Yamane <nick.yamane@openbossa.org> on 2013-11-11
Reviewed by Anders Carlsson.

After r158967, gcc debug builds with "-Werror=type-limits" enabled
fail. This patch fixes it.

* rendering/shapes/RasterShape.h:
(WebCore::RasterShapeIntervals::intervalsAt):

Canonical link: https://commits.webkit.org/142355@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159065 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/rendering/shapes/RasterShape.h
@@ -60,13 +60,13 @@ private:
 
     IntShapeIntervals& intervalsAt(int y)
     {
-        ASSERT(y + m_shapeMargin >= 0 && y + m_shapeMargin < m_intervalLists.size());
+        ASSERT(static_cast<int>(y + m_shapeMargin) >= 0 && y + m_shapeMargin < m_intervalLists.size());
         return m_intervalLists[y + m_shapeMargin];
     }
 
     const IntShapeIntervals& intervalsAt(int y) const
     {
-        ASSERT(y + m_shapeMargin >= 0 && y + m_shapeMargin < m_intervalLists.size());
+        ASSERT(static_cast<int>(y + m_shapeMargin) >= 0 && y + m_shapeMargin < m_intervalLists.size());
         return m_intervalLists[y + m_shapeMargin];
     }
